# specs/003-user-auth/contracts/auth-api.openapi.yaml
openapi: 3.0.3
info:
  title: Collab Todo - Authentication & Authorization API
  description: |
    API specification for user authentication and authorization features in the collaborative to-do application.
    
    ## Authentication
    - Uses JWT tokens stored in HTTP-only cookies
    - Session expiration: 24 hours
    - Cookie name: `auth_token`
    
    ## Authorization Hierarchy
    1. **Owner** (authenticated user who created list): Full control (read, write, delete)
    2. **Edit Link** (shareId): Read and write access (no delete)
    3. **View Link** (viewId): Read-only access
    
    ## Rate Limiting
    - Signup: 3 attempts per IP per hour
    - Signin: 5 attempts per email per 15 minutes
    
    ## Security
    - All endpoints use HTTPS in production
    - Passwords hashed with bcrypt (cost factor 12)
    - Sensitive endpoints require authentication
  version: 1.0.0
  contact:
    name: Collab Todo API Support
servers:
  - url: http://localhost:3000
    description: Local development server
  - url: https://api.collabtodo.app
    description: Production server
...
paths:
...
  /api/lists/{shareId}:
    get:
      tags:
        - Lists
      summary: Get list by shareId or viewId
      description: |
        **Authorization**: Owner OR valid edit link (shareId) OR valid view link (viewId)
        
        No authentication required if accessing via valid link.
      operationId: getList
      security:
        - cookieAuth: []
        - {}
      parameters:
        - name: shareId
          in: path
          required: true
          description: Share ID (edit link) or View ID (view-only link)
          schema:
            type: string
            pattern: '^[a-zA-Z0-9]{9}$'
            example: aBc123XyZ
      responses:
        '200':
          description: List retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  list:
                    $ref: '#/components/schemas/TodoList'
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/TodoItem'
        '404':
          description: List not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    patch:
      tags:
        - Lists
      summary: Update list metadata (e.g., title)
      description: |
        **Authorization**: Owner OR valid edit link (shareId)
        
        View-only link (viewId) cannot perform this operation.
      operationId: updateList
      security:
        - cookieAuth: []
        - {}
      parameters:
        - name: shareId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  minLength: 1
                  maxLength: 200
      responses:
        '200':
          description: List updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  list:
                    $ref: '#/components/schemas/TodoList'
        '403':
          description: Insufficient permissions (view-only access)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: Forbidden
        '404':
          description: List not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      tags:
        - Lists
      summary: Delete list (owner only)
      description: |
        **UPDATED**: Now requires authenticated owner.
        
        **Authorization**: Owner ONLY (edit/view links cannot delete)
        
        **Note**: Lists without owner (created before auth system) cannot be deleted.
      operationId: deleteList
      security:
        - cookieAuth: []
      parameters:
        - name: shareId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: List deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: List deleted successfully
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Not the owner (cannot delete via link)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: Only the list owner can delete this list
        '404':
          description: List not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/lists/{shareId}/items:
    post:
      tags:
        - Lists
      summary: Add item to list
      description: |
        **Authorization**: Owner OR valid edit link (shareId)
        
        View-only link (viewId) cannot perform this operation.
      operationId: createItem
      security:
        - cookieAuth: []
        - {}
      parameters:
        - name: shareId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - text
              properties:
                text:
                  type: string
                  minLength: 1
                  maxLength: 500
      responses:
        '201':
          description: Item created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  item:
                    $ref: '#/components/schemas/TodoItem'
        '403':
          description: Insufficient permissions (view-only or no access)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: List not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: auth_token
      description: JWT token stored in HTTP-only cookie
  schemas:
...
    TodoList:
      type: object
      required:
        - _id
        - shareId
        - viewId
        - title
        - createdAt
        - updatedAt
      properties:
        _id:
          type: string
          format: objectId
          example: 65a1b2c3d4e5f6a7b8c9d0e2
        userId:
          type: string
          format: objectId
          example: 65a1b2c3d4e5f6a7b8c9d0e1
          description: Owner user ID (optional for old lists created before auth)
        shareId:
          type: string
          pattern: '^[a-zA-Z0-9]{9}$'
          example: aBc123XyZ
          description: Edit link identifier (9-char alphanumeric)
        viewId:
          type: string
          pattern: '^[a-zA-Z0-9]{9}$'
          example: xYz987AbC
          description: View-only link identifier (9-char alphanumeric)
        title:
          type: string
          maxLength: 200
          example: Team Sprint Tasks
        createdAt:
          type: string
          format: date-time
          example: '2026-01-06T11:00:00.000Z'
        updatedAt:
          type: string
          format: date-time
          example: '2026-01-06T11:00:00.000Z'
    TodoItem:
      type: object
      required:
        - _id
        - text
        - completed
        - order
        - createdAt
        - updatedAt
      properties:
        _id:
          type: string
          format: objectId
          example: 65a1b2c3d4e5f6a7b8c9d0e3
        text:
          type: string
          maxLength: 500
          example: Write API documentation
        completed:
          type: boolean
          example: false
        order:
          type: integer
          example: 0
          description: Display order (for sorting)
        createdAt:
          type: string
          format: date-time
          example: '2026-01-06T11:05:00.000Z'
        updatedAt:
          type: string
          format: date-time
          example: '2026-01-06T11:05:00.000Z'
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          example: Invalid email format
          description: Human-readable error message